test: renaming class and init to api.py
